refactor(backend): refactor routes
