+++ backend/controller/board/board.go
@@ -1,4 +1,4 @@
-package controller
+package board
 
 import (
 	"github.com/gin-gonic/gin"
